requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
decorator==4.0.4          # via ipython, traitlets
defusedxml==0.4.1         # via python3-openid
dj-database-url==0.3.0
django-bootstrap3==6.2.2
django-extensions==1.5.7
django-postgrespool==0.3.0
django-sslify==0.2.7
django==1.8.5
djangorestframework==3.3.0
gunicorn==19.3.0
ipython-genutils==0.1.0   # via traitlets
ipython==4.0.0
oauthlib==1.0.3           # via python-social-auth, requests-oauthlib
opbeat==3.1.3
path.py==8.1.2            # via pickleshare
pexpect==4.0.1            # via ipython
pickleshare==0.5          # via ipython
psycopg2==2.6.1
ptyprocess==0.5           # via pexpect
pyjwt==1.4.0              # via python-social-auth
python-social-auth==0.2.13
python3-openid==3.0.7     # via python-social-auth
requests-oauthlib==0.5.0  # via python-social-auth
requests==2.8.1
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-extensions, python-social-auth
sqlalchemy==1.0.9         # via django-postgrespool
traitlets==4.0.0          # via ipython
whitenoise==2.0.4
